divide function added included floats

+++ src/main/java/Calculator.java
@@ -15,7 +15,7 @@ public class Calculator {
         return num1 * num2;
     }
 
-    public int divide(int num1, int num2){
+    public float divide(float num1, float num2){
         return num1 / num2;
     }
 }
